cmd/geth: blsync integrated into geth
